Add fwu standalone service context
To support service level of the fwu service, a standalone service
context is added that presents a service endpoint for fwu. This
is incorprated into the linux-pc deployment of libts. This allows
clients to discover and connect to the fwu service for the purpose
of testing. To allow different firmware configurations to be tested,
fwu components are configured outside of the fwu service context.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I4f3aa89590eab514a509172a96220b431382263d