Add fwu ref scenarios tests to ts-service-test/linux-pc
The ref_scenario tests exercise a wide range of firmware update
scenarios from the perspective of the update agent service interface
and the bootloader interface (via the fwu metadata). This change
adds the tests to the ts-service-test/linux-pc to provide service
level testing via the remote RPC based service interface. Because
the tests rely simulated device reboots, the tests are incompatible
with being run from userspace on a physical device.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I89115a17a0f9ecbf8e24b0a4c23206ed089e2f01