Add fwu configurator component
An FWU service deployment can be configured in different ways to
meet platform requirements. This commit adds a standard interface
for creating alternative configuartions with an underlying directory
structure that allows additional configurators to be added. A GPT
based configurator is included that creates the FWU configuartion
using information read from partition entries read from the GPT.
This is expected to be a common configuration method.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: Idbe8daf198bb36b8b1de340df838cc82ec450f75