Fix sp_layout.json generation
TF-A build scripts use a json configuration file (sp_layout.json) to
know which additional files need to be added to the FIP package. Each SP
will generate and install a fragment of this json file and depending on
the set of SPs to be deployed, these fragments have to be merged into a
single file.
TS has no assembly layer and thus there is no knowledge about the needed
set of SPs. As a workaround TS assumes all SPs in the install tree need
to be deployed and SP build process will merge all available json
fragments. The main benefit is decreased complexity in integration
systems.

Originally this merge was executed at cmake configuration time. This is
wrong as installation time the install directory might be different. As
a result the merged json file ends up in the binary directory of the SP
and thus remains hidden.
This change fixes the issue by moving the merging to the install step.

Signed-off-by: Gyorgy Szing <Gyorgy.Szing@arm.com>
Change-Id: Ie273c396fe81b55d94ce7f10b8cc4e9c35c0bd82