Add GPT iterator
To enable a FWU configuration to be derived from the GPT created by
a platform integrator, it is necessary to iterate over GPT entries
and to identify entries that correspond to updatable firmware
partitions. This relies on the ability to read the partition type
guid to detemine the purpose of the partition. This commit adds
an iterator that allows the raw partition table to be iterated
over to make all needed information available to support FWU
configuration.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I283788fa12d999381a2e649d6b4f8ea6aea6b124