Extend gpt block store configurator to provide GPT access
To allow block_store clients to base their own configuration on
information in the GPT, this change adds an additional storage
partition to the configuration of the partitioned_block_store
to expose the GPT contained at the head of a UEFI disk image.
This will be exploited for FWU configuration.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I59efc1d091ffa5a050b632c7382ac97320b91d93