Add volume_factory
The volume_factory defines a common interface for constructing and
destroying storage volume objects. A factory may construct any
concrete volume to suite the needs of a deployment. A single flash
volume factory is included that constructs volume objects that
present storage partitions on a block device as volumes. A single
storage device is assumed by the factory.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I12a0a070c41cbcf2a8dbddaabbbde2aba4f244fa