Change legacy-elf-format SP property to elf-format
Remove legacy- prefix from the legacy-elf-format SP manifest property
based on the comments of OP-TEE maintainers. It's unnecessary to point
out in the property name that the ELF format is legacy compared to the
recommended raw binary format.
Also adding the elf-format property to manifest files of opteesp
env-test deployments.

Signed-off-by: Imre Kis <imre.kis@arm.com>
Change-Id: I4e9cf935505697de5f4093c14db21b303735c302