Add fwu service level tests
Adds a simple fwu service-level test suite that can be run as
part of ts-service-test in a native PC and embedded environment.
Tests avoid any update activation via restart to allow tests to be
run with other service tests. This change includes making the
FWU service discoverable for FFA deployments.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I96fa7bbb2c3ae6246244080a60df5ea944a6a726