Add build option for using hardcoded IAK
The attestation SP doesn't rely on having a hardcoded attestation key,
contrary to TF-M and PSA Arch test. Because of this we had a patch for
the PSA Arch test repo to stop it from using the hardcoded key. However,
as part of the corstone1000 platform enablement to make the SE proxy SP
work with TF-M as its backend, this patch was removed, breaking the 601
test case when ran against the attestation SP. This commit reintroduces
the mentioned patch to disable using the hardcoded key again. To keep
the SE Proxy SP functional at the same time, it adds a build option that
re-enables using the hardcoded key. This option should be used when
building for the corstone1000 platform.

Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>
Change-Id: I7df92a5ae6225e52c87cb011032a18f6979ff350