Add block_store adapter for UEFI FVB driver
Adds an adapter for using UEFI FVB drivers within the block_store
framework. The adapter presents the generic block_store interface
at its upper edge and calls into the underlying FVB instance. This
is intended to allow FVB drivers from external projects such as
edk2 to be reused by trusted services.

FirmwareVolumeBlock.h is a modified version of the same file part of
the EDK2 project. The source was:
  URL: https://github.com/tianocore/edk2
  Path: MdePkg/Include/Protocol/FirmwareVolumeBlock.h
  Revision: 2f88bd3a1296c522317f1c21377876de63de5be7

This is an unfinished change. See the comment in FirmwareVolumeBlock.h.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I555bed95b5a0105d93237d36ec88a2f838de85a3