Fix includes in linux_trng_adapter
The getrandom() syscall and all of its flags are normally provided by
glibc in sys/random.h. However, it looks like there are some version
mismatches and the linux-pc deployment of component-test fails to build
in our reference build environment because GRND_INSECURE is not defined.

The Docker image in our reference build environment uses Ubuntu 18.04.6,
which has glibc 2.27. The GRND_INSECURE flag was added in Linux v5.6 and
first appeared in glibc 2.32, so the old glibc version doesn't have this
flag. However, it seems to be backported to the linux-libc-dev package
in this Ubuntu version, which provides the kernel headers. Let's include
<linux/random.h> too as a workaround, since it does have GRND_INSECURE.

Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>
Change-Id: Ie6ffb36bd76f6205131f198c6d68ad74d69a0dca