Fix access attributes in smm varibale service tests
A recent change to the uefi_variable_store corrected handling of
variable attributes being set to zero on a set_variable operation.
To be in-line with edk2, this should be interpreted as a remove
operation. The existing service tests weren't updated in-line
with this change in the previous commit, resulting in a test
failure. This commit fixes the tests.

Signed-off-by: Julian Hall <julian.hall@arm.com>
Change-Id: I53e1430c9ffb5731075454c59ff585e5f4be40e0