Fix: LazyFetch incorrectly setting build type
LazyFetch.cmake passes XXX_BUILD_TYPE as CMAKE_BUILD_TYPE to each
external component which defines a build type. The code is incorrect and
an extra indirection is missing when evaluating the value of the
variable. Thus CMAKE_BUILD_TYPE is set to the variable name instead of
it's value. This problem was not causing in any build or runtime error.

This change adds the missing indirection.

Change-Id: I12e4cd4b35225dc7cef7104cf5f5217017580a36
Signed-off-by: Gyorgy Szing <Gyorgy.Szing@arm.com>