Parse HW features from the SP manifest
Read the "arm,hw-features" node of the SP manifest and add its contents
to the config store. This way the SPMC can pass information about which
HW features are implemented by the core, since the SP itself cannot
discover this info (to read the necessary feature description registers
the exception level must be EL1 or higher).

Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>
Change-Id: I47844189c70e536817d32b674c7253f4f16efd75