Do not export symbols from libs in libts
When the libts library is linked all of the symbols that come from other
static libraries that are linked to libts are exported by default. This
behaviour is undesired, since we want to control the public interface of
libts. Also it causes problems e.g. in the linux-pc deployment where the
MbedTLS static lib is linked into libts, making all psa_* symbols
exported. Fix the issue by adding the necessary linker option.

Link: https://sourceware.org/binutils/docs-2.40/ld.html#index-_002d_002dexclude_002dlibs
Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>
Change-Id: I60fd376b274941cfc1d2ae29a2fb5634e446f6ce