Extend sp_msg register count by moving routing extension bits
Move routing extension bits to the top of W3 and make the rest of W3
available in the sp_msg structure. This way the caller of the
sp_messaging layer can use up to five argument registers. If the routing
extension is enabled the two MSBs of the first register are not
available for the caller.

Signed-off-by: Imre Kis <imre.kis@arm.com>
Change-Id: I06b155272631662adb8713aade1edbf48a3bc4a9