cmake: enhance UUID handling
Before this change SP_UUID_LE was needed as an input parameter for
export_ep(). SP_UUID_LE is used in manifest files and was set with
different methods in opteesp and sp deployments. The former used cmake
code to generate it from the SP_UUID while the latter used hand defined
values.
This change makes the two deployment types work the same way. For
this the following is done:
  - UUID manipulation routines are moved to a dedicated cmake file
  - export_sp is changed to generate the LE UUID format internally using
    the new UUID services mentioned above
  - TargetCompileDefinitions.cmake is changed to use the new UUID
    services, and to stop setting the SP_UUID_LE variable.

Change-Id: I18f4b97bb560e1ebbce3acb8aed8c3912ab05a3f
Signed-off-by: Gyorgy Szing <Gyorgy.Szing@arm.com>