Add tests for variable authentication
Creating tests to verify the new authentication feature in the
uefi service and updating the old tests to meet the changes.
Also adding tool to generate all the inputs that are needed
for the aforementioned functionality.

Moreover remove UEFi tests as these are executed part of a dedicated
test executable.

Signed-off-by: Gabor Toth <gabor.toth2@arm.com>
Signed-off-by: Gyorgy Szing <gyorgy.szing@arm.com>
Change-Id: I7aa533dcad582f7300895a15bfacdb7d2f465041