Add pkcs7 verification stub to all callers
The crypto_caller_verify_pkcs7_signature() function was only implemented
in the packed-c crypto caller. This causes an error in the SE Proxy SP,
since it contains crypto/client/psa but uses either the psa_ipc or the
stub crypto caller which do not contain this function. Add a dummy
implementation to all callers to fix this issue.

Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>
Change-Id: If3b550cae0cc1d7a62250dcf076d7dcc4f4b7886